Fixed regression in weakobjectmap no longer calling __missing__ if defined
Unlike a dict, MutableMapping doesn't automatically call __missing__ when
a key lookup fails, so it has to be explicitly called from __getitem__

A side effect of this change is that now __missing__ receives the actual
key being looked up, instead of objectid(key) as before, which is a change
for the better, as __missing__ should receive the key for which the lookup
failed. We had the objectid(key) hack before because of the particular
details of how the weakobjectmap(dict) implementation worked. While this
makes the MutableMapping based implementation backward incompatible, it is
also to be considered an inprovement.