Fixed crash when using log.capture_output()
When you use log.capture_output() it tried to encode the (input) unicode string
to the system encoding. In python3 this will give you a bytes object. The
buffer where you try to append this bytes object is a <str>, hence it fails.
Since the default iencoding for <str> in python3 is utf-8, we probably don't
need to encode it.